build folder restructure
