+++ src/cordova/setup.action.mjs
@@ -22,8 +22,8 @@ import rmfr from 'rmfr';
 import cordovaLib from 'cordova-lib';
 const {cordova} = cordovaLib;
 
-import {runAction} from '../../scripts/run_action.mjs';
-import {getCordovaBuildParameters} from '../../scripts/get_cordova_build_parameters.mjs';
+import {runAction} from '../build';
+import {getCordovaBuildParameters} from './get_cordova_build_parameters.mjs';
 
 const CORDOVA_PLATFORMS = ['ios', 'osx', 'android', 'browser'];
 const WORKING_CORDOVA_OSX_COMMIT = '07e62a53aa6a8a828fd988bc9e884c38c3495a67';
