+++ src/www/build.action.mjs
@@ -20,9 +20,7 @@ import path from 'path';
 import electronConfig from './webpack_electron.mjs';
 import cordovaConfig from './webpack_cordova.mjs';
 
-import {getBuildParameters} from '../../scripts/get_build_parameters.mjs';
-import {getBuildEnvironment} from '../../scripts/get_build_environment.mjs';
-import {getWebpackBuildMode} from '../../scripts/get_webpack_build_mode.mjs';
+import {getBuildParameters, getBuildEnvironment, getWebpackBuildMode} from '../build';
 
 /**
  * @description Builds the web UI for use across both electron and cordova.
