+++ src/www/start.action.sh
@@ -15,7 +15,7 @@
 # limitations under the License.
 set -eu
 
-node ./scripts/run_action.mjs cordova/setup browser --buildMode=debug
+node ./src/build/run_action.mjs cordova/setup browser --buildMode=debug
 
 webpack serve \
     --mode=development \
